Split integration tests into multiple runs
