+++ .travis.yml
@@ -11,6 +11,23 @@ matrix:
     sudo: true
     env:
       - INTEGRATION_TEST=true
+      # integration tests are split into multiple module sets
+      # as they take too long to run all in one execution
+      - MODULE_SET=1
+  - os: linux
+    python: 3.7
+    dist: xenial
+    sudo: true
+    env:
+      - INTEGRATION_TEST=true
+      - MODULE_SET=2
+  - os: linux
+    python: 3.7
+    dist: xenial
+    sudo: true
+    env:
+      - INTEGRATION_TEST=true
+      - MODULE_SET=3
   - os: linux
     python: 3.5
   - os: linux
